move isNearMatch to own util function
