+++ packages/gatsby/src/bootstrap/__tests__/get-config-file.ts
@@ -1,5 +1,5 @@
 import path from "path"
-import { isNearMatch, getConfigFile } from "../get-config-file"
+import { getConfigFile } from "../get-config-file"
 import { testRequireError } from "../../utils/test-require-error"
 import reporter from "gatsby-cli/lib/reporter"
 
@@ -40,20 +40,6 @@ const testRequireErrorMock = testRequireError as jest.MockedFunction<
 const reporterPanicMock = reporter.panic as jest.MockedFunction<
   typeof reporter.panic
 >
-
-describe(`isNearMatch`, () => {
-  it(`should NOT find a near match if file name is undefined`, () => {
-    const nearMatchA = isNearMatch(undefined, `gatsby-config`, 1)
-    expect(nearMatchA).toBeFalse()
-  })
-
-  it(`should calculate near matches based on distance`, () => {
-    const nearMatchA = isNearMatch(`gatsby-config`, `gatsby-conf`, 2)
-    const nearMatchB = isNearMatch(`gatsby-config`, `gatsby-configur`, 2)
-    expect(nearMatchA).toBeTrue()
-    expect(nearMatchB).toBeTrue()
-  })
-})
 
 // Separate config directories so cases can be tested separately
 const dir = path.resolve(__dirname, `../__mocks__/get-config`)
